+++ src/msg.rs
@@ -8,7 +8,7 @@ pub struct InitMsg {}
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
 pub enum HandleMsg {
-    Add { credits: i32, address: HumanAddr },
+    Add { credits: u64, address: HumanAddr },
     Reset {},
 }
 
@@ -28,12 +28,12 @@ pub struct CountResponse {
 // We define a custom struct for each query response
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct CreditsResponse {
-    pub owned_credits: i32,
+    pub owned_credits: u64,
 }
 
 // We define a custom struct for each query response
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct AddCreditsResponse {
     pub receiver: String,
-    pub new_balance: i32,
+    pub new_balance: u64,
 }
